Move email config from project.config to project.emailConfig and hide them unless specifically asked for
